power: Back up SOC to charger data register upon shutdown
Although this kernel does not read the battery's last SOC upon bootup,
nearly every other dumpling kernel does. This means that when switching
between this kernel and a different dumpling kernel (such as when booting
into a recovery image with a different kernel), there is a disparity in
the reported battery percentage since the different dumpling kernels will
read an obsolete battery percentage from the charger register.

In order to fix this, always write the last SOC to the charger register
upon shutdown. Note that we are not going to read the last SOC upon boot
because doing so would be nonsense.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Co-authored-by: Michele Beccalossi <beccalossi.michele@gmail.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>